+++ R/s2_order.R
@@ -1,5 +1,12 @@
 #' @title Order S2 products (deprecated).
 #' @description This function is deprecated and will be removed.
+#' @param s2_prodlist deprecated
+#' @param export_prodlist deprecated
+#' @param delay deprecated
+#' @param apihub deprecated
+#' @param service deprecated
+#' @param reorder deprecated
+#' @return deprecated
 #' @export
 
 s2_order <- function(